requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.8
django-extensions==1.8.1
ipdb==0.10.3
ipython==6.1.0
pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
tox==2.7.0
